requirements.txt
nltk~=3.6.5
numpy~=1.20.1
pandas~=1.2.2
pdfkit~=0.6.1
Pillow~=8.1.0
selenium~=3.141.0
setuptools~=53.0.0
py~=1.10.0
colorama~=0.4.4
toml~=0.10.2
attrs~=21.2.0
pytest~=6.2.5
pluggy~=1.0.0
iniconfig~=1.1.1
platformdirs~=2.4.0
mccabe~=0.6.1
pyparsing~=2.4.7
opencv-python-headless~=4.5.4.58
wheel~=0.37.0
editdistance~=0.6.0
dacite
